LayoutTests/imported/w3c/web-platform-tests/css/css-viewport/zoom/relative-units-expected.txt


FAIL relative-units assert_approx_equals: em in outside expected 20 +/- 1 but got 10
FAIL relative-units 1 assert_approx_equals: rem in outside expected 20 +/- 1 but got 10
FAIL relative-units 2 assert_approx_equals: lh in outside expected 20 +/- 1 but got 10
FAIL relative-units 3 assert_approx_equals: rlh in outside expected 20 +/- 1 but got 10
FAIL relative-units 4 assert_approx_equals: vh in outside expected 12 +/- 1 but got 3
FAIL relative-units 5 assert_approx_equals: vw in outside expected 16 +/- 1 but got 4
